% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contents.R
\name{contents}
\alias{contents}
\title{description of a data frame}
\usage{
contents(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a list with 4 components:
\describe{
\item(dfname)(name of data frame)
}
}
\description{
\code(contents) provides describes the content
of a data frame
}
\details{
fou each variable in a data frame, \code(contents) describs
the position,type,number of unique values ,
}
\examples{
df_info <- contents(happiness)
df_info
plot(df_info)
}
\seealso{
\link(print.contents), \link(plot.contents)
}
