% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.contents.R
\name{plot.contents}
\alias{plot.contents}
\title{Visualize a data frame}
\usage{
\method{plot}{contents}(x, ...)
}
\arguments{
\item{x}{an object of class \code(contents)}
}
\value{
a \code(ggplot2) graph
}
\description{
\code(plot.contents)
}
\details{
for each variable ,the plot displays
\itemize{
\item(type(\code(numeric),
\code(factor)))
\item(percent of available cases)
}
}
\examples{
df_info <- contents(happiness)
plot(df_info)
}
\seealso{
see \link(contents).
}
