% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.contents.R
\name{print.contents}
\alias{print.contents}
\title{description of a data frame}
\usage{
\method{print}{contents}(x, digits, ...)
}
\arguments{
\item{x}{an object of class \code(contents)}
}
\description{
\code(print.contents) prints a concise description of a data frame
}
\examples{
df_info <- contents(happiness)
print(df_info,digits=4)
}
